+++ src/messages/protocol.ts
@@ -63,6 +63,7 @@ export type CommandType =
   | "commitAndPush"
   | "amendCommit"
   | "rollbackFile"
+  | "rollbackFiles"
   | "getWorkingTreeChanges"
   | "getShelves"
   | "shelveChanges"